show CBR label for CBR enforced builds
